mamake: require(): fix Mam/shell syntax conflict (re: 07b240d4)
In require(), which may be run by the Mam 'bind' command, mamake
may execute a test program to check if a library exists. It puts
a shell script in a buffer and then runs the buffer through
expand() to expand any Mam variables in it. Thing is, Mam
expansions start with `${` and in certain contexts they may
conflict with shell expansions which also may start with `${`.
The following is valid Mam syntax:

    ${TMPDIR:-/tmp}

which expands to empty. It is yet to be documented exactly what the
full syntax of Mam expansions is; the expand() code need to be
analysed. Yet, it turns out that

    ${TMPDIR-/tmp}

does not conflict and gets passed to the shell unmodified.
It is good enough to use that instead. It allows for empty values
of a set TMPDIR to be used by the shell but that is pathological
anyway. If TMPDIR is entirely unset, it still does the right thing.