mamake: allow simple 'done'
This takes another small step towards making the Mamfiles easier to
read and maintain by hand: repeating the make target after 'done'
is now optional. If it is given, it is still used as a sanity check
and any mismatch with 'make' will produce an error. This is handy
for long or all-encompassing rules. But for short ones, the
repetitions were basically just noise; indentation suffices.