Revert "Mamfiles: remove a large amount of redundant cruft"
This reverts commit ce6c1252d61b222c8a2e9fc7ef985ae30f33549f.

That commit was based on a serious reasoning error. Just because
the libast files are known to exist doesn't mean they are known to
be up to date with ksh and other directories that use libast. For
instance, after that commit, if something changes in ast_map.h
(e.g. singal being renamed to _ast_signal via a #define) and ksh
isn't recompiled due to the deleted prev statements, then ksh will
fail to link.