Add SHOPT_PRINTF_LEGACY: non-compliant printf option parsing
On some systems, the external printf(1) allows formatters that
begin with a '-' without a preceding '--' options terminator
argument. The acceptance of this obsolete syntax violates POSIX
because printf falls under the regular utility syntax guidelines.

This situation makes it difficult to add options to printf as an
extension (which POSIX allows) while maintaining backward
compatibility.

But some systems come with old sh scripts that depend on the
obsolete printf behaviour, so ksh needs to adapt if it is to be
accpeptable as a replacement system shell on these.

Illumos, in particular, ships a patched ksh 93u+ 2012-08-01 that
completely disables option parsing in printf and replaces it with a
kludge that only recognises the POSIX '--' options terminator.
This matches some external printf commands but kills 'printf --man'
and friends and makes the addition of 'printf -v' impossible.

This commit adds a compile-time option that provides good-enough
compatibility with the obsolete printf behaviour, while keeping our
own built-in printf options usable.

By default, this is only compiled in if the external printf(1)
command on the system where ksh is built exhibits the obsolete
behaviour. It can be manually (de)activated regardless of this by
setting SHOPT_PRINTF_LEGACY to 0 or 1 src/cmd/ksh93/SHOPT.sh.

src/cmd/ksh93/SHOPT.sh,
src/cmd/ksh93/Mamfile:
- Add SHOPT_PRINTF_LEGACY compile-time option.
- When no value is given, probe if external printf(1) accepts a
  formatter starting with - without preceding --, and emit a
  #define for SHOPT_PRINTF_LEGACY if found.

src/cmd/ksh93/bltins/print.c: b_print():
- If SHOPT_PRINTF_LEGACY is on and b_print() is called from printf,
  maximise backward compatibility without killing our options using
  a two-pronged approach:
  - If the first argument contains a % or a \, disable option
    parsing: that is very likely a format operand even if it starts
    with '-', and it cannot be a valid printf option.
  - When about to throw an error on an invalid option (c==':'),
    check if the error occurred on the first argument (note that
    optget leaves opt_info.index==1 on error in a short option, but
    2 on error in a long option). If so, ignore the error and treat
    the argument as a format operand regardless of content.
    (Note that for --man, etc., c=='?', which is not affected.)

src/cmd/ksh93/sh/shtests:
- For the regression tests, we need to check of SHOPT_PRINTF_LEGACY
  was enabled by default at compile time, which we cannot tell from
  the SHOPT.sh file. Scan the generated shopt.h file and use its
  definitions as variables as well, overriding any from SHOPT.sh.

src/cmd/ksh93/tests/builtins.sh:
- Do not check for correct handling of an invalid option in printf
  if SHOPT_PRINTF_LEGACY is on.