Properly fix ignored SIGINT on interactive (re: 55dc80ce, 7e5fd3e9)
The kludge added in the first referenced commit was one I found
experimentally without understanding it properly at the time.

It was necessary because the Red Hat patch kept the subsequent
siglongjmp() after changing the sh_fault() call to killpg().

But changing it to killpg() made the siglongjmp inappropriate
either way: it is not reached if SIGINT is handled, and it should
not be executed if SIGINT is ignored.

So, that kludge was actually a compensation for the latter case.

src/cmd/ksh93/edit/edit.c:
- Remove the siglongjmp() call so that ed_getchar() simply
  continues as if nothing happened if the killpg() call returns.
  This means the interrupt character is now ignored properly along
  with SIGINT and longer causes a new prompt to be started.
     (If the interrupt character hasn't been changed from the
  default Ctrl+C using 'stty intr', then in emacs, the Ctrl+C
  feature (capitalise current character) is exposed as a result.)

src/cmd/ksh93/sh/main.c:
- Remove the kludge added to exfile().