fun/man: act correctly on 'man builtin | editor'
When the regular man(1) output is piped into another command, then
man(1) adapts: it disables terminal formatting and sets the column
width to 80. This function should do the same for the built-in
commands so that e.g. 'man iffe | vi' works as expected and does
not show a lot of control codes in the editor.