sh_exec: don't fail to longjmp (re: d309d604)
In d446015e, we learned that it can cause problems to fail to
longjmp after sigsetjmp, for sufficiently high values of jmpval.
In this case, after nv_setlist, we should longjmp when jmpval is
higher than SH_JMPCMD (error in a command).

I don't know of any current regressions this may fix, but it's more
correct this way.