Fix libast/Mamfile dependencies; mamake strict lvl 3 (re: 4eaf25c8)
It turns out I introduced a number of dependency bugs while
refactoring src/lib/libast/Mamfile: a number of dependencies were
declared with 'prev' on targets that were not yet made. As a
result, the dependency information in the subsequent 'make' blocks
was ignored.

There were also a number of circular dependencies that we inherited
from the AT&T-generated Mamfiles; these are now fixed as well, or
at least deactivated and marked. Previously, they were ignored.

All except one of the circular dependencies actually represent
circular #include dependecies in AST headers, something that we may
need to work out at some point.

To avoid this reoccurring, we need to tighten things up more. This
commit introduces a new strict level 3, in which:

* Making a prerequisite that is currently being made, or one that
  has already been made, is an error.

* Declaring a dependency (with 'prev') on a prerequisite that is
  currently being made is an error.

At strict levels < 3, we now throw a warning in these cases.

The Mamfiles are fixed to use and comply with the new strict level,
and should be more correct as a result. The only one that needed
significant changes was src/lib/libast/Mamfile.