Fix handling of env vars in namespaces (re: 0ea4e559)
This commit fixes corner case breakage of univese handling in the
'echo' command: it breaks when used in a namespace block before
being used anywhere else, and only if _AST_FEATURES is not exported
first. The bug causes _AST_FEATURES to be incorrectly placed in
that namespace.

  $ (unset _AST_FEATURES; ksh -c 'namespace foo { echo test; }')
  ksh: echo: _AST_FEATURES=UNIVERSE - ucb: is not an identifier

The specific code change in the referenced commit that triggers the
bug is the change from dtinsert to dtinstall in nv_search(). Change
that back to dtinsert and the reproducer works again.

src/lib/libast/man/cdt.3 says:
|  OBJECT OPERATIONS
|    void* dtinsert(Dt_t* dt, void* obj)
|    void* dtinstall(Dt_t* dt, void* obj)
|    void* dtappend(Dt_t* dt, void* obj)
|      These functions add an object prototyped by obj into dt.
|      See Dtdisc_t.makef for object construction. dtinsert() and
|      dtappend() perform the same function for all methods except
|      for Dtlist (see Dtlist for details). For Dtset, Dtrhset or
|      Dtoset, if there is an object in dt matching obj, dtinsert()
|      and dtappend() will not insert a new object, whereas
|      dtinstall() will remove such a matching object before
|      inserting the new object.

I don't know exactly why this change from dtinsert to dtinstall
triggers the bug, but I don't think there is anything wrong with
it; the node was not found, a new one is created, so if there is an
old matching node after all (for whatever reason), it *should* be
removed or we have a memory leak.

The real problem, exposed by that commit, is in the shell intercept
functions of getenv(3), putenv(3) and setenviron(3): they fail to
unset the namespace. libast calls AST setenviron(3) to set the
_AST_FEATURES variable. (See src/lib/libast/man/setenviron.3) Those
functions are intercepted so that environment variables set by
libast immediately become available as shell variables.

src/cmd/ksh93/sh/name.c:
- Fix the getenv(3), putenv(3) and setenviron(3) shell overrides
  to not only temporarily deactivate the prefix for compound
  assignment, but to do the same for the active namespace as well.

Thanks to @stephane-chazelas for the report.
Resolves: https://github.com/ksh93/ksh/issues/727