package: DEFPATH fix for Android/Termux (re: 0a0a32c3)
The standard utilities directory on Android is actually /bin (a
symlink to /system/bin). There lives sh (which is dash) and the
core utilities (from Toybox). In DEFPATH these should take
precedence over whatever is added by termux -- if only because /bin
is read-only and therefore more secure; termux utilities are
installed by and can be overwrittem by the regular user.