Revert dynamic linking of mamake for now (re: 2e849e67)
It somehow conflicts with System Integrity Protection on macOS
13.6.6 (even though it's fine on macOS 12.6.7). Since I have no way
to test it on 13.6.6 and find a workaround, I'm simply reverting it
for now. Dynamically linking mamake against libast is useful as a
test case for dynamic linking, but non-essential.

src/lib/libast/Mamfile:
- Revert.

bin/package:
- When rebuilding mamake, remove both versionslinked agaisnt libast
  (static and dymamic) to avoid them found in PATH.

src/cmd/INIT/mamake.c:
- Version date bump to trigger rebuild (and thus removal of
  dynamically linked version as per above change).

Resolves: https://github.com/ksh93/ksh/issues/735