Fix: optional builtin not found when run from namespace
This bug only shows up if /opt/ast/bin is not in PATH.

Reproducer:

    PATH=$(getconf PATH) /path/to/ksh -c \
	'builtin basename; namespace ns { basename --version; }'

Output:

    basename: illegal option -- -
    usage: basename string [suffix]
           basename [-a] [-s suffix] string [...]

Expected output:

      version         basename (ksh 93u+m) 2022-08-30

The bug only occurs if the namespace block is on the same line as
the preceding builtin basename command. Inserting a newline fixes
it. So, the issue is related to line-by-line parsing.

src/cmd/ksh93/sh/xec.c: sh_exec():
- Not only do the fallback search for a path-bound built-in for
  built-ins invoked by full canonical paths, but also for simple
  names. This catches this corner case and possibly others.

Resolves: https://github.com/ksh93/ksh/issues/728