build: trigger redo of mamprobe after editing C+probe or make.probe
In f2bc1f45, I wrote:

> src/cmd/INIT/mamprobe.sh:
> - Update version number. This change causes mamake to redo the
>   probe (it checks for a change in mamprobe but not in make.probe;
>   I should probably fix that at some point).

Normally, mamprobe is updated directly by bin/package. This fixes
the problem by adding a rule in the Mamfile for it where it is not
just dependent on the mamprobe.sh source but also on probe.sh,
which is the combination of C+probe and make.probe. This will then
cause mamake to redo the compiler info probe when one of the latter
two is edited, as well.