conf.tab: Support //UNC/paths by default on Cygwin
In the Cygwin port for ksh 93u+m/1.0.8, Roland Mainz had patched
astconf(3) to hardcode a value 1 for PATH_LEADING_SLASHES, which
is necessary to enable the built-in support for the double initial
slash used for UNC network paths on Windows.

Previously, enabling this support required manually adding the
value triplet "PATH_LEADING_SLASHES - 1" to the _AST_FEATURES
environment variable, but the documentation for that was so obscure
and indirect it might as well have been undocuemnted (it would
require reading the --man self-doc of the getconf, dirname and
basename path-bound built-ins).

This commit adds a PATH_LEADING_SLASHES entry to conf.tab that will
cause this value to be hardcoded to 1 in the generated conftab.c
file on Cygwin, causing UNC path support to always be enabled on
that system.